Refactor:#76 - standardize schedule icon naming, location, and format
- Relocate from public/assets/icons/scheduleIcon.svg to src/assets/icons/schedule.svg
- Apply consistent SVGR props
